pyproject.toml
[build-system]
requires = ["setuptools >= 67.7.2"]
build-backend = "setuptools.build_meta"

[project]
name = "openant"
description = "ANT, ANT-FS and ANT+ Python Library"
authors = [
    {name = "Gustav Tiger", email = "gustav@tiger.name"},
    {name = "John Whittington"}
]
requires-python = ">=3.8"
license = {file = "LICENSE"}
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Embedded Systems",
    "Environment :: Console"
]
dependencies = [
    "pyusb>=1.1.1"
]
dynamic = ["readme", "version"]

[project.optional-dependencies]
serial = ["pyserial"]
influx = ["influxdb-client"]
mqtt = ["paho-mqtt"]
test = ["pytest", "black", "pylint"]
docs = ["sphinx>=5.2.3", "furo>=2021.3.20b30", "sphinx_mdinclude"]

[project.scripts]
openant = "openant.__init__:_main"

# specify to avoid symbolic link ant
[tool.setuptools.packages.find]
include = ["openant*"]

[tool.setuptools.dynamic]
readme = { file = "README.md" }
version = { attr = "openant.__version__" }

[tool.setuptools.package-data]
openant = ["resources/42-ant-usb-sticks.rules", "examples/*.py"]

[tool.black]
line-length = 88
target-version = ['py37', 'py38', 'py39', 'py310', 'py311', 'py312']

[tool.pylint]
disable = [
    "C0103", # (invalid-name)
    "C0114", # (missing-module-docstring)
    "C0115", # (missing-class-docstring)
    "C0116", # (missing-function-docstring)
    "R0903", # (too-few-public-methods)
    "R0913", # (too-many-arguments)
    "R0801", # (similar-lines)
    "W1203", # (logging-fstring-interpolation)
    "W0107", # (pass-statement)
    "R0902",
    "C0301", # (line-too-long)
]
[tool.pylint.LOGGING]
logging-format-style = "new"

[tool.pytest.ini_options]
addopts = "--verbose --doctest-modules --ignore=openant/subparsers/influx.py --ignore=openant/subparsers/mqtt.py"
doctest_optionflags = "NORMALIZE_WHITESPACE IGNORE_EXCEPTION_DETAIL"
testpaths = "openant"
